move the metrics stuff to its own file
